requirements.txt
Flask==1.0.3
simplejson==3.16.0
google-cloud-datastore==1.8.0
